Add user roles

+++ todo_list/views.py
@@ -6,10 +6,15 @@ from rest_framework.decorators import api_view
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.authtoken.models import Token
 from django.shortcuts import get_object_or_404
-from django.contrib.auth.models import User
+from django.contrib.auth.models import User, Group, Permission
 from django.contrib.auth import authenticate
+
+from permissions import IsAdminOrReadOnly
 from .serializers import ToDoListSerializer, UserSerializer
 from .models import ToDoList
+
+admin, _ = Group.objects.get_or_create(name='admin')
+user, _ = Group.objects.get_or_create(name='user')
 
 class ToDoListViewSet(viewsets.ModelViewSet):
     queryset = ToDoList.objects.all()
@@ -23,7 +28,7 @@ class ToDoListViewSet(viewsets.ModelViewSet):
         serializer.save(user=self.request.user)
     
 class ToDoListView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsAdminOrReadOnly]
     
     def get(self, request):
         todos = ToDoList.objects.all()
@@ -31,6 +36,9 @@ class ToDoListView(APIView):
         return Response(serializer.data)
     
     def post(self, request):
+        if not request.user.groups.filter(name='admin').exists():
+            return Response({"error": "You do not have permission to create todos."}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = ToDoListSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -38,7 +46,7 @@ class ToDoListView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
 class ToDoDetailView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsAdminOrReadOnly]
     
     def get(self, request, pk):
         todo = get_object_or_404(ToDoList, pk=pk)
@@ -47,6 +55,10 @@ class ToDoDetailView(APIView):
 
     def put(self, request, pk):
         todo = get_object_or_404(ToDoList, pk=pk)
+        
+        if not request.user.groups.filter(name='admin').exists():
+            return Response({"error": "You do not have permission to create todos."}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = ToDoListSerializer(todo, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -55,6 +67,11 @@ class ToDoDetailView(APIView):
 
     def delete(self, request, pk):
         todo = get_object_or_404(ToDoList, pk=pk)
+        
+        if not request.user.groups.filter(name='admin').exists():
+            return Response({"error": "You do not have permission to create todos."}, status=status.HTTP_403_FORBIDDEN)
+
+        
         todo.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
@@ -64,6 +81,16 @@ class RegisterView(generics.CreateAPIView):
     serializer_class = UserSerializer
     permission_classes = [permissions.AllowAny]
     
+    def perform_create(self, serializer):
+        user = serializer.save()
+        role = self.request.data.get('role','user')
+        if role == 'admin':
+            admin_group = Group.objects.get(name='admin')
+            user.groups.add(admin_group)
+        else:
+            user_group = Group.objects.get(name='user')
+            user.groups.add(user_group)  
+              
 class LoginView(APIView):
     permission_classes = [permissions.AllowAny]
     
@@ -77,5 +104,7 @@ class LoginView(APIView):
             return Response({"error": "Invalid credentials"}, status=status.HTTP_400_BAD_REQUEST)
 
         token, created = Token.objects.get_or_create(user=user)
+        
+        role = 'admin' if user.groups.filter(name='admin').exists() else 'user'
         
         return Response({"token": token.key}, status=status.HTTP_200_OK)
